Add PlayerStats

+++ Assets/Scripts/EnemyStats.cs
@@ -3,10 +3,10 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "Enemy", menuName = "ScriptableObjects/Enemy")]
 public class EnemyStats : ScriptableObject
 {
-    public int damage;
+    public float damage;
     public float attackRange;
     public float moveSpeed;
-    public int maxHealth;
+    public float maxHealth;
     public bool isBoss;
 
     public int exp;
